pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langsmith>=0.4.47",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "tavily-python>=0.7.13",
]
